requirements.txt
Django==2.1.2
py-moneyed==0.7.0
pytz==2018.5
djangorestframework==3.8.2